[FIX] tools.translate: when loading entries from a PO file, use also the POT.
On Launchpad, as commented on lp:933496, PO entries (andtheir comments) are
shared between series. This mean that e.g. the 6.1 series can have the wrong
`reference` comments (those beginning with #:) as they would be copied from say
the trunk series.  Those `reference` comments are used by OpenERP to create its
translations and look them up.

This patch adds a few lines of code to tools.translate.py so that targets
defined in the POT `reference` comments are used in addition to those from the
PO file.