[FIX] move loaded templates storage to context
Didn't think of the _t attribute storing loaded templates when qweb
was moved from a Python object instantiated every rendering to an
openerp never reinstantiated. This unexpectedly moved _t from the
status of a per-render cache (to avoid reloading sub-templates or
recursive templates a billion times) into a permanent cache with no
invalidation.

Move _t to QWebContext.templates, which should be per-render, for now.