[IMP] Improved reporting of test failures
While loading modules with the tests option, it's useful to have a report that
display test problems separately from the main log. It's then up to the
launching system (for instance openerpcommand) to produce meaningful user
output.
            
This would ease I hope the maintainers task in case of regressions by letting
them dispatch easily to people that happen to know the relevant modules
(together with the author of the commit that triggered the error)
            
A corresponding patch for openerpcommand will be uploaded