[FIX] res.font: re-introduce substitution for builtin PDF fonts
This substitution was removed when res.font was implemented
based on the idea that it would be best to allow the use of
the real Helvetica/Times/Courier font if really required.
However the builtin ones are always available and crippled
when it comes to Unicode supports. Combined with the fact
that many RML/SXW reports had Helvetica hardcoded, this
breaks unicode reports for many users.
This patch re-introduces a mapping to replace the builtin
fonts by a few common and free alternatives (DejaVu, Liberation
and Free font families).
When an actual Helvetica/Times/Courier font is installed
no substitution takes place for it, in the hope that it
comes with decent unicode support.